+++ bighomework-1/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=kpo
+spring.application.name=bighomework-1
